Add `bInvalidHandleIsAcceptable` argument to `fbWaitForSingleObject`
If provided and set to `True`, waiting for an invalid handle will not throw an error but return `True` instead. Used for Process/Thread handles, which may be invalid because the Process/Thread no longer exists, in which case the handle is invalid but one commonly expects the function to return `True` to indicate the Process/Thread has terminated.